requirements.txt
Django>=5.0.0
djangorestframework
django-cors-headers
django-filter
python-dotenv
djangorestframework-simplejwt
celery
redis
whitenoise
gunicorn
dj-database-url
psycopg2-binary
